refactor(mobile): use hook
