+++ apps/mobile/src/modules/Home/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useMap';
